// src/types/exam.ts
export interface Question extends Omit<IQuestion, '_id' | 'createdAt' | 'updatedAt'> {
  _id: string;
  createdAt: string;
  updatedAt: string;
  correctOptionsCount?: number;
  totalOptionsCount?: number;
}